dbr quant function fusion [1/x]: record matches for functions
Summary:

For DBR quant, adds the code for matching seen ops to function fusion
patterns. After we have the full DAG, we have a separate pass over the
dag and add matched fusion patterns to the seen op data structure.

This is the first PR in the stack which implements matching and
recording the match results. Future PRs in this stack will use
the match results to modify observer insertion and inference.

Test plan:

```
python test/test_quantization.py TestQuantizeDBR.test_fusion_functions
```

ghstack-source-id: 3f7ed238a08e55e25add6e21f4af49a45be59ffd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71764