dbr quant function fusion [2/x]: use fusion for observation and inference
Summary:

The previous PR added information about fusions found in the subgraphs.

This PR uses that information for:
1. inserting observers at the end of fusions and not in the middle
2. during inference, replacing the original op with the fused op. The
way this is implemented is that the base op is replaced with the fused op,
and all other ops are replaced with identity functions.

Test plan:

```
python test/test_quantization.py TestQuantizeDBR.test_fusion_functions
```

ghstack-source-id: e0cf9378c2b526162eda325eeac43546d67e4ba7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71781